+++ src/transceiver/base_station/base_station_to_user_equipment_link.py
@@ -1,0 +1,6 @@
+from link import Link
+
+
+class BaseStationToUserEquipmentLink(Link):
+    def __init__(self, source_node, destination_node, channel):
+        super().__init__(source_node, destination_node, channel)
