+++ src/transceiver/user_equipment/user_equipment_to_base_station_link.py
@@ -1,0 +1,5 @@
+from link import Link
+
+class UserEquipmentToBaseStationLink(Link):
+    def __init__(self, source_node, destination_node, channel):
+        super().__init__(source_node, destination_node, channel)
