--- src/userequipmenttobasestationlink.py
@@ -1,5 +1,0 @@
-from link import Link
-
-class UserEquipmentToBaseStationLink(Link):
-    def __init__(self, source_node, destination_node, channel):
-        super().__init__(source_node, destination_node, channel)
