+++ test/test_userequipmentcreator.py
@@ -3,8 +3,8 @@
 import unittest
 # sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
-from src.userequipmentcreator import UserEquipmentCreator
-from src.userequipment import UserEquipment
+from transceiver.user_equipment.user_equipment_creator import UserEquipmentCreator
+from transceiver.user_equipment.user_equipment import UserEquipment
 
 class TestUserEquipmentCreator(unittest.TestCase):
